feat: add support for parsing complex command lines
This diff adds support for parsing complex command lines
featuring several subcommands.

We reuse the simple parsing funcionality.

Complex parsing functionality is a wrapper around it.

We basically build the command line model manually using
a declarative syntax where reading the code gives a visual
hint of how the command line usage would look like.

(At least, I hope 😬.)

Then, the game basically consists on parsing until we
reach a leaf node in tree of parsers.

When this happens we wrap the remaining command line arguments,
subject to constraints on their number and return.

We understand which leaf node was selected by switching on
the type of the returned option structure.